Project 1 from semester 2
Allocate Arrays to Hold the Data
The first line in the file says how many data records there are. You will be using a
Scanner to process the file. Read that number and then create 5 separate integer arrays
of that size for month, day, year, tmax, and tmin. These will hold all the data from
the file. This is a lot of data, but your computer can easily hold it in memory.
2. Read and Store the Data in your Arrays
Next call your Scanner’s nextLine() method 3 times to discard the remainder of the
first line (”data records”) and the next two lines (”date tmax tmin” and the horizonal
line below them). Next you are going to read and store all the data into your arrays.
The Scanner’s default behavior is to break up data at whitespace (spaces, tabs, and
newlines), but we also want to break apart the date string which uses forward slashes
to separate the month, day and year. Use this code to change the delimiter to alsoinclude forward slashes before starting to read the actual data:
fileScnr.useDelimiter("[/ \t\n\r]+");
Now you can just use .nextInt() to read each of the five fields: month, day, year, tmax,
and tmin. The Scanner will then automatically wrap to the next line and you can
continue reading the entire file, and storing it in your arrays.
Verify that you have read in the data correctly. Either examine your arrays using the
debugger or print them. Do they match the contents of the file?3. Long Term Averages and Records
(a) Write a method that finds and returns the index of the maximum value in a 1-D
array that is passed to it.
(b) Write a method that finds and returns the index of the minimum value in 1-D
array that is passed to it.
(c) Write a method that finds and returns the average (a double) of the numbers in
a 1-D array that is passed to it.
(d) Write a method that finds and returns the first index of a given value in a 1-D
array between two indexes that are passed to it.
Use those methods to find and print:
 The highest temperature in tmax and the date it occurred on and then use that
index for your month[], day[], and year[] arrays)
 The lowest temperature in tmin and the date it occurred on
 The average tmax (Check your value. It should be 55.6608...)
 The average tmin
4. Finding the starting index for each year.
One aspect of climate change that has been noted in Maine is that the first fall frost
has been occurring later in the year. We are going to analyze the minimum daily
temperature data to calculate the date each fall when the temperature first reaches
freezing temperatures. First we are going to find the starting index of each year in the
’year’ array. Your task is to create a loop that prints:
year starting index
1941 0
1942 365
1943 730
1944 1095
1945 1461Remember it is poor programming to hard code specific values such as 1941 and 2020
in a program because then the program would have to be modified to process a different
file. Instead use year[0] to get the starting year and year[n-1] to get the last year. (I
used the variable n for the number of data records.) I suggest that you store these
values in a 1-D array. You should be able to figure out how long to make the array.
You know how to access the first year and the last value and there are 12 months in
each year. Although there might be missing data, you can still make a good guess at
the length of the needed array since it is the number of years.
5. Finding the first fall frost
Make a method
int firstFallFrost(int[] tmin, int yearStartIndex, int yearEndIndex)
that looks at a one year section of the array for the date of the first fall frost and
returns the index of that day. Hint: start searching forward from the middle of the
year. For testing, you should find that for 1941 (tmin, 0, 365) it occurs at index 262,
and for 1942 (tmin, 365, 730) it occurs at index 636. In the next part you will use the
month and day arrays to print the human understandable dates 9/20 and 9/29.
6. Printing the first fall frost dates
Use your starting year indices from Part 4, your firstFallFrost() method from part 5,
and the month and day arrays, to print a table of the first fall frost dates for each year
like this:
year first fall frost
1941 9/20
1942 9/29
1943 10/10
1944 9/25
1945 9/30
Make sure your results match.
7. Averaging by decade:
A nicer way to present the data is averaged by decade, i.e. the average fall frost day
in the 1940’s, 1950’s, ... 1920’s. Write code to do the analyses this way. Note that
for some decades we may not have data for all 10 years (in our data the 1940’s and
the 2020’s). Your code should still give correct averages for these decades. Print your
results in a neat table. Use printf and format specifiers NOT tabs or spaces to line
up your columns.
8. Analysis of the results:
The first fall frost dates vary a lot from year to year. Compare the average of the first
10 years of the historical record to the average of the most recent 10 years. How much
has it changed? Print a simple statement about the change in the average first fall
frost date over these 80 years. The following method should be added to your program
to do this analysis: